+++ CHANGES.md
@@ -5,9 +5,8 @@
 ## 1.2.0
 
 - Fixed running tests on 2021.3 platform version
-- Delay dependencies resolving to tasks configuration phase
-- Deprecate `IntelliJPluginExtension.getIdeaDependency(project: Project)`
 - Avoid downloading IDE dependency in configuration phase
+- Deprecate `IntelliJPluginExtension.getIdeaDependency(project: Project)`
 - Increase the default `runPluginVerifier.failureLevel` to `COMPATIBILITY_PROBLEMS`
 - Introduce `listProductsReleases` task for listing the IDE releases matching given criteria
 - Fixed resolving compiler classpath for the `instrumentCode` task when using `LATEST-EAP-SNAPSHOT` [#752](../../issues/752)