+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## next
 
 - Respect `ideaDependencyCachePath` property [#794](../../issues/794)
+- Fix for providing dependencies after project evaluation [#801](../../issues/801)
 
 ## 1.2.0