+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## next
+## 1.2.1
 
 - Respect `ideaDependencyCachePath` property [#794](../../issues/794)
 - Fix for providing dependencies after project evaluation [#801](../../issues/801)